refactor(core): move required to schema and improve types
Moved the required information from property to the schema.
This needed to be done, because a schema can exists that has properties in the required list from one of the anyOf or allOf sub schemas.
This could not have been represented by the boolean on the schema property.
Also replace some array with Sets and Maps to ensure uniqueness and hopefully improve performance a bit.